+++ source/Databricks/Databricks.sln
@@ -15,7 +15,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "release-notes", "release-no
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "source", "source", "{C3C7D8A2-E32D-4353-A70F-5D88D52E48D3}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SqlStatement", "SqlStatement\SqlStatement.csproj", "{86DF757A-943C-42D7-87B5-2F71495AC5F5}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SqlStatement", "SqlStatement\SqlStatement.csproj", "{86DF757A-943C-42D7-87B5-2F71495AC5F5}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -33,7 +33,6 @@ Global
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{65F2891F-9158-4EE1-8296-D44C5388D5C7} = {A70FC42B-F0A0-43EC-B76E-7586811FCF13}
-		{C3C7D8A2-E32D-4353-A70F-5D88D52E48D3} = {A70FC42B-F0A0-43EC-B76E-7586811FCF13}
 		{86DF757A-943C-42D7-87B5-2F71495AC5F5} = {C3C7D8A2-E32D-4353-A70F-5D88D52E48D3}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
