+++ source/Databricks/SqlStatement/AppSettings/DatabricksOptions.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Energinet.DataHub.Core.Databricks.SqlStatement;
+namespace Energinet.DataHub.Core.Databricks.SqlStatement.AppSettings;
 
 public class DatabricksOptions
 {
